# Generalized surreal numbers parametrized by option sorts