--- a/contrib/HoTTBook.v
+++ b/contrib/HoTTBook.v
@@ -2283,8 +2283,7 @@ Definition Book_10_5_8_item_ix := @HoTT.HIT.V.separation.
 (* ================================================== defn:surreals *)
 (** Definition 11.6.1 *)
 
-Definition Book_11_6_1 := @HoTT.Spaces.No.Core.Surreals.No.
-Definition Book_11_6_1 := @HoTT.Spaces.No.Core.MaxNo.
+Definition Book_11_6_1 := @HoTT.Spaces.No.Core.No.
 
 (* ================================================== thm:NO-simplicity *)
 (** Theorem 11.6.2 *)