--- a/theories/Spaces/No/Addition.v
+++ b/theories/Spaces/No/Addition.v
@@ -65,18 +65,18 @@ Section Addition.
           intros [l|l] [r|r]; cbn;
           [ apply xL_lt_xR_plus
           | transitivity ((xL_plus l).1 (yR r));
-            [ apply (snd (xL_plus l).2), Conway_theorem0_ii_r; exact _
+            [ apply (snd (xL_plus l).2), lt_ropt; exact _
             | exact (fst (x_plus_yR r).2 l) ]
           | transitivity ((xR_plus r).1 (yL l));
             [ exact (snd (x_plus_yL l).2 r)
-            | apply (snd (xR_plus r).2), Conway_theorem0_ii_l; exact _ ]
+            | apply (snd (xR_plus r).2), lt_lopt; exact _ ]
           | apply x_plus_yL_lt_yR ]). }
         assert (InSort S L'' R'') by (apply sum_options; exact _).
         exists ({{ zL | zR // zcut }}); split.
         + intros l.
-          refine (Conway_theorem0_ii_l zL zR zcut (inl l)).
+          refine (lt_lopt zL zR zcut (inl l)).
         + intros r.
-          refine (Conway_theorem0_ii_r zL zR zcut (inl r)).
+          refine (lt_ropt zL zR zcut (inl r)).
       - abstract (
         intros x y [a ?] [b ?] p q r s;
         rewrite transport_sigma; cbn in *;
@@ -90,11 +90,11 @@ Section Addition.
         apply le_lr; [ intros [l|l] | intros [r|r] ]; cbn;
         [ refine (le_lt_trans (fst (xL_plus l).2 _ {{ zL | zR // zcut}} _) _);
           [ by (apply le_lr; assumption)
-          | refine (Conway_theorem0_ii_l _ _ _ (inl l)) ]
+          | refine (lt_lopt _ _ _ (inl l)) ]
         | exact (x_plus_yL_lt_z l)
         | refine (lt_le_trans _
                     (fst (xR_plus r).2 {{ yL | yR // ycut}} _ _));
-          [ refine (Conway_theorem0_ii_r _ _ _ (inl r))
+          [ refine (lt_ropt _ _ _ (inl r))
           | by (apply le_lr; assumption) ]
         | exact (x_plus_y_lt_zR r) ] ).
       - abstract (
@@ -178,14 +178,14 @@ Section Addition.
         exact xL_lt_y_plus
       | (** x + z^L < y + z *)
         refine (le_lt_trans (x_le_y_plus_zL l) _);
-        refine (Conway_theorem0_ii_l _ _ _ (inr l))
+        refine (lt_lopt _ _ _ (inr l))
       | (** x + z < y^R + z *)
         specialize (x_lt_yR_plus r {{ zL | zR // zcut }});
         rewrite p in x_lt_yR_plus;
         exact x_lt_yR_plus
       | (** x + z < y + z^R *)
         refine (lt_le_trans _ (x_le_y_plus_zR r));
-        refine (Conway_theorem0_ii_r _ _ _ (inr r)) ]).
+        refine (lt_ropt _ _ _ (inr r)) ]).
     - abstract (
       intros L R ? xL xR xcut xL_plus xR_plus xL_lt_xR_plus
              L' R' ? yL yR ycut yL_plus yR_plus yL_lt_yR_plus;
@@ -198,7 +198,7 @@ Section Addition.
       destruct (plus_inner_cut yL_plus yR_plus yL_lt_yR_plus
                                zL zR zcut) as [yzcut q];rewrite q;
       refine (le_lt_trans (x_le_yL_plus {{ zL | zR // zcut }}) _);
-      refine (Conway_theorem0_ii_l _ _ _ (inl l)) ).
+      refine (lt_lopt _ _ _ (inl l)) ).
     - abstract (
       intros L R ? xL xR xcut xL_plus xR_plus xL_lt_xR_plus
              L' R' ? yL yR ycut yL_plus yR_plus yL_lt_yR_plus;
@@ -211,7 +211,7 @@ Section Addition.
       destruct (plus_inner_cut yL_plus yR_plus yL_lt_yR_plus
                                zL zR zcut) as [yzcut q];rewrite q;
       refine (lt_le_trans _ (xR_le_y_plus {{ zL | zR // zcut }}));
-      refine (Conway_theorem0_ii_r _ _ _ (inl r)) ).
+      refine (lt_ropt _ _ _ (inl r)) ).
   Defined.
 
   Definition plus (x y : No) : No