--- a/theories/Spaces/No/Addition.v
+++ b/theories/Spaces/No/Addition.v
@@ -16,6 +16,21 @@ Class HasAddition (S : OptionSort) :=
 Existing Instance empty_options.
 Existing Instance sum_options.
 
+Global Instance hasaddition_maxsort : HasAddition MaxSort
+  := { empty_options := tt ;
+       sum_options := fun _ _ _ _ _ _ => tt }.
+  
+Global Instance hasaddition_ordsort : HasAddition OrdSort
+  := { empty_options := idmap ;
+       sum_options := fun _ _ _ _ f g => sum_ind _ f g }.
+
+Global Instance hasaddition_decsort : HasAddition DecSort.
+Proof.
+  constructor.
+  - apply insort_decsort.
+  - intros L R L' R' [? ?] [? ?]; split; exact _.
+Qed.
+
 Section Addition.
   Context `{Univalence}.
   Context {S : OptionSort@{i}} `{HasAddition S}.