… theories/Spaces/No/Negation.v