Basically, you can get a "notion of surreal numbers" by restricting the types parametrizing the left and right options in any way you like, and the basic theorems about order are completely unaffected, while addition and so on do just fine as long as the option sorts are closed under the appropriate operations.  In particular, this includes the plump ordinals from Exercise 11.17 in the book.  All kinds of generalized surreal numbers embed into the "maximal" surreal numbers, though I don't know of a good way to define their images intrinsically.  Finally, and interestingly, *every* surreal number is equal to one whose option sorts are decidable (i.e. either it has no left options or it has a left option, and either it has no right options or it has a right option).