@mikeshulman 2017-07-28
Oh, and the timing diff for this entire changeset starts out

```
After    | File Name                                                 | Before   || Change   
--------------------------------------------------------------------------------------------
4m27.16s | Total                                                     | 4m25.59s || +0m01.56s
--------------------------------------------------------------------------------------------
0m05.67s | Spaces/No/Core                                            | 0m03.72s || +0m01.94s
0m03.00s | Spaces/No/Addition                                        | 0m01.50s || +0m01.50s
```

and then degenerates into noise.  So the extra typeclass inference is not adding hardly any noticeable overhead.