@mikeshulman 2017-07-28
I don't really understand the timing diffs for this.  Here's the diff for be830d17db, which I would expect to have all the effect of the extra typeclass inference:

```
After    | File Name                                                 | Before   || Change   
--------------------------------------------------------------------------------------------
4m25.55s | Total                                                     | 4m26.11s || -0m00.55s
--------------------------------------------------------------------------------------------
0m02.98s | Spaces/No/Addition                                        | 0m01.52s || +0m01.46s
0m32.84s | Spaces/BAut/Bool                                          | 0m33.66s || -0m00.81s
0m18.80s | Categories/Category/Sigma/Univalent                       | 0m19.15s || -0m00.34s
0m11.27s | Algebra/ooGroup                                           | 0m11.41s || -0m00.14s
0m10.24s | Comodalities/CoreflectiveSubuniverse                      | 0m10.50s || -0m00.25s
0m08.78s | contrib/Freudenthal                                       | 0m08.88s || -0m00.10s
0m08.05s | Modalities/ReflectiveSubuniverse                          | 0m07.89s || +0m00.16s
```

It seems that the typeclass inference is not slowing down `No/Core` but it is slowing down `No/Addition`... and yet there is enough timing noise in totally unrelated files to cancel out that slowdown?