@mikeshulman 2017-07-28
Hmm, okay, running it again the noise goes the other way, but the slowdown in `No/Addition` remains constant, plus a more minor slowdown in `No/Core` is also noticeable:

```
After    | File Name                                                 | Before   || Change   
--------------------------------------------------------------------------------------------
4m28.05s | Total                                                     | 4m25.90s || +0m02.14s
--------------------------------------------------------------------------------------------
0m03.02s | Spaces/No/Addition                                        | 0m01.52s || +0m01.50s
0m33.41s | Spaces/BAut/Bool                                          | 0m33.65s || -0m00.24s
0m19.06s | Categories/Category/Sigma/Univalent                       | 0m19.08s || -0m00.01s
0m11.35s | Algebra/ooGroup                                           | 0m11.33s || +0m00.01s
0m10.48s | Comodalities/CoreflectiveSubuniverse                      | 0m10.56s || -0m00.08s
0m08.88s | contrib/Freudenthal                                       | 0m09.02s || -0m00.13s
0m07.79s | Modalities/ReflectiveSubuniverse                          | 0m07.96s || -0m00.16s
0m07.29s | Pointed                                                   | 0m07.36s || -0m00.07s
0m07.24s | Spaces/BAut/Rigid                                         | 0m07.25s || -0m00.00s
0m07.20s | Spaces/Finite                                             | 0m07.20s || +0m00.00s
0m06.48s | Spaces/BAut                                               | 0m06.48s || +0m00.00s
0m05.75s | Categories/Adjoint/Pointwise                              | 0m05.78s || -0m00.03s
0m05.68s | Idempotents                                               | 0m05.65s || +0m00.02s
0m05.51s | HIT/Colimits/Flattening                                   | 0m05.50s || +0m00.00s
0m05.04s | Modalities/Lex                                            | 0m04.96s || +0m00.08s
0m04.19s | Modalities/Topological                                    | 0m04.17s || +0m00.02s
0m04.12s | Categories/Adjoint/Functorial/Laws                        | 0m04.02s || +0m00.10s
0m04.00s | Spaces/No/Core                                            | 0m03.68s || +0m00.31s
```

Suggestions for speedup are welcome, though it's only a couple seconds in total.