@SkySkimmer 2017-07-28
On my machine specifically the `Definition plus_cut` goes from 0.4s to 1.7s.
I don't think it's because of typeclasses, `Typeclasses eauto := debug.` only shows
```
Debug: 1: looking for (InSort S L R) with backtracking
Debug: 1.1: exact Sx on (InSort S L R), 0 subgoal(s)
Debug: 2: looking for (InSort S L' R') with backtracking
Debug: 2.1: exact Sy on (InSort S L' R'), 0 subgoal(s)
Debug: 3: looking for (HasAddition S) without backtracking
Debug: 3.1: exact H0 on (HasAddition S), 0 subgoal(s)
Debug: 4: looking for (InSort S L R) without backtracking
Debug: 4.1: exact Sx on (InSort S L R), 0 subgoal(s)
Debug: 5: looking for (InSort S L' R') without backtracking
Debug: 5.1: exact Sy on (InSort S L' R'), 0 subgoal(s)
plus_cut is defined
Finished transaction in 1.668 secs (1.624u,0.008s) (successful)
```