@mikeshulman 2017-07-28
Well, it may not be literally spending the time in typeclass inference, but the only thing that changed in that commit was the use of a typeclass, so it must somehow be indirectly connected, no?