@mikeshulman 2017-08-05
If no one has any speedup suggestions, maybe we could go ahead and merge this and put "speed it up" as a to-do item for later?