@JasonGross 2017-08-11
> Hmm, okay, running it again the noise goes the other way,

Note that there are `-only-diff` versions of the timing scripts, which exclude files that don't depend on the ones that you changed, and therefore have less noise in the overall time.

I'm looking into the timing change now.