@mikeshulman 2017-08-11
Thanks.  (FWIW I think the timing scripts could be better documented, and maybe have an easier interface to use and remember; I can never remember which script is which or the name of the variable `PREV_COMMIT` or whatever it is.)