@JasonGross 2017-08-11
> FWIW I think the timing scripts could be better documented, and maybe have an easier interface to use and remember

I've updated the `pretty-timed-*diff-tip.sh` files' documentation at the top to mention `PREV_COMMIT`.  Further concrete suggestions and contributions welcome.  (I'm not entirely sure what a better interface would be, or what better documentation would look like.)