@JasonGross 2017-08-11
> On my machine specifically the `Definition plus_cut` goes from 0.4s to 1.7s.

Inferring the argument `Sx` to `plus_inner_cut` takes around 0.6 to 1.0 seconds.  Adding `(Sx:=Sx)` to `plus_inner_cut` in `plus_cut` drops the time from 1.445 s to 0.391 s on my machine.

Regarding `No/Core.v`, the line `apply IHL0, No_decode_le.` takes 0.8 seconds.  If you replace it with `apply IHL0, (@No_decode_le _).`, it takes only 0.003 seconds.  I found this line by using `etc/coq-scripts/timing/make-both-single-timing-files.py` on the `theories/Spaces/No/Core.timing` files created by running make with each of the two states I was comparing checked out.

I think that should take care of most of your timing differences.