@mikeshulman 2017-08-11
Thanks!  I implemented those changes.